refactor: cli doesn't need to depend on deno_tls (#12952)
also move create_http_client to deno_fetch
