+++ cli/Cargo.toml
@@ -45,7 +45,6 @@ deno_doc = "0.21.0"
 deno_graph = "0.12.0"
 deno_lint = { version = "0.19.0", features = ["docs"] }
 deno_runtime = { version = "0.35.0", path = "../runtime" }
-deno_tls = { version = "0.14.0", path = "../ext/tls" }
 
 atty = "0.2.14"
 base64 = "0.13.0"
